requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20
pydantic==2.11.7
python-dotenv==1.1.1
requests==2.32.4
groq==0.10.0

